solving cgal build on mac
