+++ OpenMEEG/libs/OpenMEEGMaths/include/matrix.h
@@ -188,8 +188,7 @@ namespace OpenMEEG {
     inline double Matrix::frobenius_norm() const {
     #ifdef HAVE_LAPACK
     if ( nlin()*ncol() != 0 ) {
-        double work = -1.;
-        work += 1.; // <- just to remove the warning unused-variable ...
+        double work;
         return DLANGE('F', sizet_to_int(nlin()), sizet_to_int(ncol()), data(), sizet_to_int(nlin()), &work);
     } else {
         return 0;
