+++ OpenMEEG/tools/cgal_lib_function.h
@@ -37,31 +37,31 @@ The fact that you are presently reading this means that you have had
 knowledge of the CeCILL-B license and that you accept its terms.
 */
 
+#pragma once
+
+// for verbosity
+#define CGAL_MESH_3_VERBOSE
+
 #include <mesh.h>
-#include "options.h"
-#include <cgal_lib.h>
 
-using namespace OpenMEEG;
+#include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
+#include <CGAL/Mesh_3/Robust_intersection_traits_3.h>
+#include <CGAL/Polyhedron_3.h>
+#include <CGAL/config.h>
+#include <CGAL/version.h>
 
-int main(int argc, char **argv) {
-    command_usage("Create a BEM mesh from either an implicit function: sphere, hemisphere, ...:");
-    const double sphere_radius     = command_option("-r", 0., "radius of the sphere");
-    const double hemisphere_radius = command_option("-hr", 0., "radius of the hemisphere");
-    const double radius_bound      = command_option("-fs",1e-1,"facet radius bound of elements");
-    const double distance_bound    = command_option("-fd",1e-1,"facet distance bound to the input surface");
-    // const unsigned init_points  = command_option("-ip", 10, "initial number of points (for the hemisphere)");
-    const char * output_filename   = command_option("-o",(const char *) NULL,"Output Mesh");
+typedef CGAL::Exact_predicates_inexact_constructions_kernel K;
+typedef K::Point_3 Point_3;
+typedef K::FT FT;
+typedef CGAL::Polyhedron_3<K>  Polyhedron;
+typedef Polyhedron::HalfedgeDS HDS;
 
-    if ( command_option("-h",(const char *)0,0) ) {
-        return 0;
-    }
-    if ( output_filename == NULL ) {
-        std::cerr << "Set an output filename" << std::endl;
-        return 0;
+namespace OpenMEEG {
+
+    namespace {
+        template <typename C3t3>
+        Mesh CGAL_to_OM(C3t3 c3t3);
     }
 
-    Mesh m_out = cgal_mesh_function(sphere_radius, hemisphere_radius, radius_bound, distance_bound);
-    m_out.save(output_filename);
-    m_out.info();
-    return 0;
+    Mesh cgal_mesh_function(double sphere_radius, double hemisphere, double radius_bound, double distance_bound);
 }
