+++ OpenMEEG/tools/cgal_mesh_3Dlabeled_image.cpp
@@ -37,11 +37,11 @@ The fact that you are presently reading this means that you have had
 knowledge of the CeCILL-B license and that you accept its terms.
 */
 
-#include <mesh.h>
 #include <options.h>
-#include "cgal_mesh.h"
+#include <cgal_lib.h>
 
 using namespace OpenMEEG;
+
 int main(int argc, char **argv) {
     command_usage("Create a BEM mesh from a 3D labeled image (e.g a mask):");
     const char * input_filename  = command_option("-i",(const char *) NULL,"Input image");
